+++ gadgets/checkTools/ui.py
@@ -31,9 +31,9 @@ class ResultLister(QtWidgets.QTreeView):
         super(ResultLister, self).__init__(parent)
         self.setHeaderHidden(True)
         self.setRootIsDecorated(False)
-        self.setEditTriggers(self.NoEditTriggers)
-        self.setVerticalScrollMode(self.ScrollPerPixel)
-        self.setHorizontalScrollMode(self.ScrollPerPixel)
+        self.setEditTriggers(QtWidgets.QTreeView.NoEditTriggers)
+        self.setVerticalScrollMode(QtWidgets.QTreeView.ScrollPerPixel)
+        self.setHorizontalScrollMode(QtWidgets.QTreeView.ScrollPerPixel)
         self.clicked.connect(self.emit_selection_changed)
         self.setExpandsOnDoubleClick(False)
         model = QtGui.QStandardItemModel(0, 1)
