+++ gadgets/objectStatsManager.py
@@ -65,11 +65,11 @@ class TargetViewer(QtWidgets.QGroupBox):
         self.__view.setModel(model)
         self.__view.setSelectionModel(sel_model)
         self.__view.setRootIsDecorated(False)
-        self.__view.setEditTriggers(self.__view.NoEditTriggers)
-        self.__view.setVerticalScrollMode(self.__view.ScrollPerPixel)
-        self.__view.setHorizontalScrollMode(self.__view.ScrollPerPixel)
+        self.__view.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
+        self.__view.setVerticalScrollMode(QtWidgets.QTreeView.ScrollPerPixel)
+        self.__view.setHorizontalScrollMode(QtWidgets.QTreeView.ScrollPerPixel)
         self.__view.setIconSize(QtCore.QSize(32, 32))
-        self.__view.setSelectionMode(self.__view.ExtendedSelection)
+        self.__view.setSelectionMode(QtWidgets.QTreeView.ExtendedSelection)
         
         editor = TargetEditor()
         # editor.reloadButtonClicked.connect(self.updateGui)
